refactor and reuse some existing functionality; extract writeChangeTypes into separate file; and make TODO tests
