+++ packages/cli/src/commands/add/changeTypes.ts
@@ -1,22 +1,15 @@
-import { info, log } from "@changesets/logger";
+import { log } from "@changesets/logger";
 import {
   ChangeType,
   ChangesetWithConfirmed,
   EmptyString,
   Release,
   Config,
-  VersionType,
-  Changeset
+  VersionType
 } from "@changesets/types";
-import path from "path";
-import { Package } from "@manypkg/get-packages";
 import chalk from "chalk";
 
-import { getCommitFunctions } from "../../commit/getCommitFunctions";
 import * as cli from "../../utils/cli-utilities";
-import printConfirmationMessage from "./messages";
-import * as git from "@changesets/git";
-import { determineEditorHack, warnIfMajor } from ".";
 import changeTypeList from "./changeTypeList.json";
 
 const { bold, cyan } = chalk;
@@ -25,7 +18,6 @@ type ChangeTypeOption = { title: string; text: string };
 const allChangeTypes: ChangeTypeOption[] = changeTypeList;
 
 type PreviousAnswers = { [key in string]: string };
-
 export class ChangesetsWithChangeTypes {
   private config: Config;
   private isWithChangeTypes: boolean = false;
@@ -191,65 +183,6 @@ async function getReleasesPerPackage(
   return changesetList;
 }
 
-type WriteChangesetListArgs = {
-  changesets: ChangesetWithConfirmed[];
-  packages: Package[];
-  cwd: string;
-  changesetBase: string;
-  empty?: boolean;
-  config: Config;
-  open?: boolean;
-  writeChangeset: (changeset: Changeset, cwd: string) => Promise<string>;
-};
-export async function writeChangesetList({
-  changesets,
-  packages,
-  cwd,
-  changesetBase,
-  empty,
-  config,
-  open,
-  writeChangeset
-}: WriteChangesetListArgs) {
-  for (let changeset of changesets) {
-    printConfirmationMessage(changeset, packages.length > 1);
-
-    if (!changeset.confirmed) {
-      changeset = {
-        ...changeset,
-        confirmed: await cli.askConfirm("Is this your desired changeset?")
-      };
-    }
-
-    if (!changeset.confirmed) continue;
-
-    const changesetID = await writeChangeset(changeset, cwd);
-
-    const [{ getAddMessage }, commitOpts] = getCommitFunctions(
-      config.commit,
-      cwd
-    );
-    if (getAddMessage) {
-      await git.add(path.resolve(changesetBase, `${changesetID}.md`), cwd);
-      await git.commit(await getAddMessage(changeset, commitOpts), cwd);
-      log(chalk.green(`${empty ? "Empty " : ""}Changeset added and committed`));
-    } else {
-      log(
-        chalk.green(
-          `${empty ? "Empty " : ""}Changeset added! - you can now commit it\n`
-        )
-      );
-    }
-
-    warnIfMajor(changeset);
-
-    const changesetPath = path.resolve(changesetBase, `${changesetID}.md`);
-    info(chalk.blue(changesetPath));
-
-    if (open) determineEditorHack(changesetPath);
-  }
-}
-
 async function getDescriptionWithPreviousAnswer(
   previousAnswers: PreviousAnswers,
   category: ChangeTypeOption
@@ -266,6 +199,7 @@ async function getDescriptionWithPreviousAnswer(
   return description;
 }
 
+// @TODO merge duplicate setSummary with existing
 async function setSummary(changeSet: ChangesetWithConfirmed, config: Config) {
   log(
     "Please enter a summary for this change (this will be in the changelogs)."
@@ -299,71 +233,4 @@ async function setSummary(changeSet: ChangesetWithConfirmed, config: Config) {
 
   changeSet.summary = summary;
   changeSet.confirmed = false;
-}
-
-function groupByBumpType(releases: Release[]) {
-  const major: Release[] = [];
-  const minor: Release[] = [];
-  const patch: Release[] = [];
-  const none: Release[] = [];
-
-  releases.forEach(rel => {
-    if (rel.type === "major") major.push(rel);
-    else if (rel.type === "minor") minor.push(rel);
-    else if (rel.type === "patch") patch.push(rel);
-    else major.push(rel);
-  });
-  return { major, minor, patch, none };
-}
-function getReleasesSection(releases: Release[]) {
-  return `---
-${releases.map(release => `"${release.name}": ${release.type}`).join("\n")}
----\n`;
-}
-function getChangeTypesSection(releases: Release[], bumpType?: VersionType) {
-  let changeTypes: ChangeType[] = [];
-
-  if (bumpType) {
-    const [oneRelease] = releases.filter(({ type }) => type === bumpType);
-    if (!oneRelease?.changeTypes) changeTypes = [];
-    else changeTypes = oneRelease.changeTypes;
-  } else {
-    changeTypes = releases.flatMap(rel => rel.changeTypes || []);
-  }
-
-  return `${changeTypes
-    .filter(chk => chk.description)
-    .map(chk => `- [ ${chk.category.title} ] ${chk.description}`)
-    .join("\n")}\n`;
-}
-
-export function getChangesetContent(
-  releases: Release[],
-  summary: string,
-  splitReleasesByBumpType = false
-) {
-  if (splitReleasesByBumpType && releases.some(rel => rel.changeTypes)) {
-    const grouped = Object.entries(groupByBumpType(releases)).filter(
-      ([, releases]) => releases.length
-    ) as [VersionType, Release[]][];
-
-    return `${grouped
-      .map(
-        ([bumpType, releases]) =>
-          `${getReleasesSection(releases)}
-${getChangeTypesSection(releases, bumpType)}`
-      )
-      .join("\n")}
-
-${summary}
-`;
-  }
-
-  if (releases.some(rel => rel.changeTypes))
-    return `${getReleasesSection(releases)} 
-  ${getChangeTypesSection(releases)}
-
-${summary}
-`;
-  return null;
 }
