+++ packages/cli/src/commands/add/index.ts
@@ -5,7 +5,7 @@ import { spawn } from "child_process";
 import * as cli from "../../utils/cli-utilities";
 import * as git from "@changesets/git";
 import { info, log, warn } from "@changesets/logger";
-import { Config } from "@changesets/types";
+import { Changeset, Config } from "@changesets/types";
 import { getPackages } from "@manypkg/get-packages";
 import writeChangeset from "@changesets/write";
 
@@ -13,11 +13,44 @@ import { getCommitFunctions } from "../../commit/getCommitFunctions";
 import createChangeset from "./createChangeset";
 import printConfirmationMessage from "./messages";
 import { ExternalEditor } from "external-editor";
-import { writeChangesetList } from "./changeTypes";
+import { writeChangesetList } from "./writeChangeTypes";
 
 type UnwrapPromise<T extends Promise<any>> = T extends Promise<infer R>
   ? R
   : never;
+
+export function warnIfMajor(changeset: Changeset) {
+  let hasMajorChange = [...changeset.releases].find(c => c.type === "major");
+
+  if (hasMajorChange) {
+    warn(
+      "This Changeset includes a major change and we STRONGLY recommend adding more information to the changeset:"
+    );
+    warn("WHAT the breaking change is");
+    warn("WHY the change was made");
+    warn("HOW a consumer should update their code");
+  } else {
+    log(
+      chalk.green(
+        "If you want to modify or expand on the changeset summary, you can find it here"
+      )
+    );
+  }
+}
+
+export function determineEditorHack(changesetPath: string) {
+  // this is really a hack to reuse the logic embedded in `external-editor` related to determining the editor
+  const externalEditor = new ExternalEditor();
+  externalEditor.cleanup();
+  spawn(
+    externalEditor.editor.bin,
+    externalEditor.editor.args.concat([changesetPath]),
+    {
+      detached: true,
+      stdio: "inherit"
+    }
+  );
+}
 
 export default async function add(
   cwd: string,
@@ -91,39 +124,10 @@ export default async function add(
       );
     }
 
-    let hasMajorChange = [...newChangeset.releases].find(
-      c => c.type === "major"
-    );
-
-    if (hasMajorChange) {
-      warn(
-        "This Changeset includes a major change and we STRONGLY recommend adding more information to the changeset:"
-      );
-      warn("WHAT the breaking change is");
-      warn("WHY the change was made");
-      warn("HOW a consumer should update their code");
-    } else {
-      log(
-        chalk.green(
-          "If you want to modify or expand on the changeset summary, you can find it here"
-        )
-      );
-    }
+    warnIfMajor(newChangeset);
     const changesetPath = path.resolve(changesetBase, `${changesetID}.md`);
     info(chalk.blue(changesetPath));
 
-    if (open) {
-      // this is really a hack to reuse the logic embedded in `external-editor` related to determining the editor
-      const externalEditor = new ExternalEditor();
-      externalEditor.cleanup();
-      spawn(
-        externalEditor.editor.bin,
-        externalEditor.editor.args.concat([changesetPath]),
-        {
-          detached: true,
-          stdio: "inherit"
-        }
-      );
-    }
+    if (open) determineEditorHack(changesetPath);
   }
 }
